+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/HistoricQueriesRepository.java
@@ -3,5 +3,5 @@ package io.pakland.mdas.githubstats.domain.repository;
 import io.pakland.mdas.githubstats.domain.HistoricQueries;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistoricQueriesRepository extends JpaRepository<HistoricQueries, Long> {
+public interface HistoricQueriesRepository extends JpaRepository<HistoricQueries, Integer> {
 }
