+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/UserReviewRepository.java
@@ -9,5 +9,5 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface UserReviewRepository extends JpaRepository<UserReview,Long> {
+public interface UserReviewRepository extends JpaRepository<UserReview,Integer> {
 }
